Corrigindo erro no calculo das metricas com 8bits.

+++ dissertacao/unet/test_unet_metrics_mixed_quant.py
@@ -61,7 +61,7 @@ def load_patient(image):
     npyImage = np.expand_dims(npyImage, axis=-1)
     npyImage = (npyImage>0.5)*1
 
-    return npyImage
+    return npyImage.astype(np.float32)
 
 
 def predictPatient(model, image):
@@ -80,8 +80,8 @@ def predictPatient(model, image):
     npyImagePredict = np.around(npyImagePredict, decimals=0)
     npyImagePredict = (npyImagePredict>0.5)*1
 
-    # return npyImagePredict.astype(np.float32)
-    return npyImagePredict
+    return npyImagePredict.astype(np.float32)
+    # return npyImagePredict
 
 
 def execPredict(exam_id, input_path, input_mask_path, output_path, model):
